Fix function export for local dev server
