+++ examples/draft-pad/lib/utils.ts
@@ -4,3 +4,16 @@ import { twMerge } from "tailwind-merge";
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
+
+export const shouldUseLocalDevServer = () => {
+  if (
+    process.env.GENSX_BASE_URL &&
+    !process.env.GENSX_BASE_URL.includes("localhost")
+  ) {
+    return false;
+  }
+  if (process.env.NODE_ENV === "production" || process.env.VERCEL_ENV) {
+    return false;
+  }
+  return false;
+};
